debugger-agi_latest/requirements.txt
Backtesting==0.3.3
bokeh==3.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
clipboard==0.0.4
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
MouseInfo==0.1.3
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
ta==0.11.0
tornado==6.4
tzdata==2023.3
urllib3==2.1.0
xyzservices==2023.10.1
youtube-transcript-api==0.6.1
